test(integration): a pushed fragment reaching a peer over iroh
`client_authored` pins what `/sync` owes a client on one node. What the
node then owes the rest of the network had no coverage, and the gap it
was hiding -- a merged fragment that never left the node it was pushed
to -- was invisible to a single-node suite by construction.

Two nodes over iroh, connected with a replicator from the first to the
second, and a fragment signed by a key neither node holds pushed to the
first. The peer is never pushed to and never asks, so the only thing
that can put the document on it is the node that merged it announcing
what it merged. The second test is the half that makes the first safe:
the peer must hold the client's signature, not one belonging to the node
that carried it.

Both fail without the announcement -- the document does not arrive
within the P2P timeout.

The fragment builders move to `client_authored_common` so the iroh
suite can reach them, following `manage_relay_common`. No test logic
changes in the move.
