+++ tools/integration-test/tests/p2p_iroh/main.rs
@@ -1,4 +1,6 @@
 mod acp;
+#[path = "../client_authored_common.rs"]
+mod client_authored_common;
 mod connection;
 #[path = "../manage_relay_common.rs"]
 mod manage_relay_common;
