+++ tools/integration-test/tests/p2p_iroh/replication/mod.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::module_inception)]
 
+mod client_authored;
 mod collection_sub;
 mod document;
 mod document_sub;
